refactor 4/4
